requirements.txt
openai>=1.0.0
requests>=2.25.0
Pillow>=8.0.0
uuid
PyYAML>=6.0
